// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  CUSTOMER: 'CUSTOMER',
  WORKER: 'WORKER',
  ADMIN: 'ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const WorkerStatus = {
  PENDING: 'PENDING',
  VERIFIED: 'VERIFIED',
  REJECTED: 'REJECTED',
  SUSPENDED: 'SUSPENDED'
} as const

export type WorkerStatus = (typeof WorkerStatus)[keyof typeof WorkerStatus]


export const BookingStatus = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
  EXPIRED: 'EXPIRED',
  DISPUTED: 'DISPUTED'
} as const

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus]


export const BookingType = {
  INSTANT: 'INSTANT',
  SCHEDULED: 'SCHEDULED',
  EMERGENCY: 'EMERGENCY'
} as const

export type BookingType = (typeof BookingType)[keyof typeof BookingType]


export const DisputeStatus = {
  OPEN: 'OPEN',
  UNDER_REVIEW: 'UNDER_REVIEW',
  RESOLVED_CUSTOMER: 'RESOLVED_CUSTOMER',
  RESOLVED_WORKER: 'RESOLVED_WORKER',
  DISMISSED: 'DISMISSED'
} as const

export type DisputeStatus = (typeof DisputeStatus)[keyof typeof DisputeStatus]


export const DisputeReason = {
  WORKER_DID_NOT_ARRIVE: 'WORKER_DID_NOT_ARRIVE',
  CUSTOMER_DID_NOT_PAY: 'CUSTOMER_DID_NOT_PAY',
  WRONG_SERVICE_PROVIDED: 'WRONG_SERVICE_PROVIDED',
  FRAUD: 'FRAUD',
  PROPERTY_DAMAGE: 'PROPERTY_DAMAGE',
  MISCONDUCT: 'MISCONDUCT'
} as const

export type DisputeReason = (typeof DisputeReason)[keyof typeof DisputeReason]


export const PaymentMethodType = {
  EASYPAISA: 'EASYPAISA',
  JAZZCASH: 'JAZZCASH',
  SADAPAY: 'SADAPAY',
  NAYAPAY: 'NAYAPAY',
  BANK_ACCOUNT: 'BANK_ACCOUNT'
} as const

export type PaymentMethodType = (typeof PaymentMethodType)[keyof typeof PaymentMethodType]


export const NotificationType = {
  BOOKING_UPDATE: 'BOOKING_UPDATE',
  CHAT_MESSAGE: 'CHAT_MESSAGE',
  PROMOTION: 'PROMOTION',
  DISPUTE_UPDATE: 'DISPUTE_UPDATE',
  VERIFICATION_UPDATE: 'VERIFICATION_UPDATE'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]


export const Language = {
  ENGLISH: 'ENGLISH',
  URDU: 'URDU',
  PUNJABI: 'PUNJABI',
  SINDHI: 'SINDHI',
  PASHTO: 'PASHTO',
  OTHER: 'OTHER'
} as const

export type Language = (typeof Language)[keyof typeof Language]